#36 - Rename lib to src, sort structure, split files.
