+++ src/Baseline.FluentHttpExtensions.Tests/Unit/HeaderExtensionsTests/WithBearerTokenAuthTests.cs
