+++ src/Baseline.FluentHttpExtensions.Tests/Unit/HeaderExtensionsTests/WithHeaderTests.cs
