+++ src/Baseline.FluentHttpExtensions.Tests/Unit/HeaderExtensionsTests/WithUserAgentTests.cs
@@ -6,12 +6,12 @@ namespace Baseline.FluentHttpExtensions.Tests.Unit.HeaderExtensionsTests
     public class WithUserAgentTests : UnitTest
     {
         [Fact]
-        public async Task User_Agent_Is_Set_In_MoogieHttpRequest_Object()
+        public async Task User_Agent_Is_Set_In_HttpRequest_Object()
         {
-            OnRequestMade(r => Assert.Equal("moogie-http", r.Headers.UserAgent.ToString()));
+            OnRequestMade(r => Assert.Equal("baseline", r.Headers.UserAgent.ToString()));
 
             await HttpRequest
-                .WithUserAgent("moogie-http")
+                .WithUserAgent("baseline")
                 .AsAGetRequest()
                 .EnsureSuccessStatusCode();
         }
