+++ src/Baseline.FluentHttpExtensions.Tests/Unit/UrlExtensionsTests/WithPathSegmentTests.cs
