+++ src/Baseline.FluentHttpExtensions/Baseline.FluentHttpExtensions.csproj
@@ -15,28 +15,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <Compile Remove="FluentHttpExtensions.Tests\**" />
-        <Compile Include="..\HttpRequest.cs">
-            <Link>HttpRequest.cs</Link>
-        </Compile>
-        <Compile Remove="Baseline.FluentHttpExtensions.Tests\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <EmbeddedResource Remove="FluentHttpExtensions.Tests\**" />
-        <EmbeddedResource Remove="Baseline.FluentHttpExtensions.Tests\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <None Remove="FluentHttpExtensions.Tests\**" />
-        <None Remove=".editorconfig" />
-        <None Remove=".gitignore" />
-        <None Remove="LICENSE" />
-        <None Remove="README.md" />
-        <None Remove="Baseline.FluentHttpExtensions.Tests\**" />
-    </ItemGroup>
-
-    <ItemGroup>
         <PackageReference Include="System.Text.Json" Version="5.0.1" />
     </ItemGroup>
